+++ server/src/SSO/SsoServer.hpp
@@ -107,7 +107,7 @@ bool cSSOServer::OnConnect(cNetworkConnection* pConnection)
         std::cout << "service with Service-ID: '" << sServiceId << "', hasn't been registered yet." << std::endl;
     }
     string sResponse = oResponse.Serialize();
-    pConnection->SendBytes((byte*)sResponse.c_str(), sResponse.size());
+    pConnection->SendBytes((byte*)sResponse.c_str(), (int) sResponse.size());
 
     pConnection->UnLockRecieve();
     return oResponse.GetResponseCode() == 200;
@@ -225,7 +225,7 @@ bool cSSOServer::HandleSessionRequest(cNetworkConnection *pConnection, cUri & oU
             string sResponse = oResponse.Serialize();
             for (auto&[sServiceId, pServiceConnection] : paServices)
             {
-                pServiceConnection->SendBytes((byte *) sResponse.c_str(), sResponse.size());
+                pServiceConnection->SendBytes((byte *) sResponse.c_str(), (int) sResponse.size());
             }
             return true;
         }
@@ -236,7 +236,7 @@ bool cSSOServer::HandleSessionRequest(cNetworkConnection *pConnection, cUri & oU
     oResponse.SetHeaders(aHeaders);
 
     string sResponse = oResponse.Serialize();
-    pConnection->SendBytes((byte*)sResponse.c_str(), sResponse.size());
+    pConnection->SendBytes((byte*)sResponse.c_str(), (int) sResponse.size());
     return true;
 }
